Changed some architecture
